Reorganize the unit tests
