+++ src/models/unit_tests/rule_test.rs
@@ -1,3 +1,5 @@
+use tree_sitter::Point;
+
 /*
 Copyright (c) 2023 Uber Technologies, Inc.
 
@@ -12,7 +14,12 @@ Copyright (c) 2023 Uber Technologies, Inc.
 */
 use crate::{
   filter,
-  models::{default_configs::UNUSED_CODE_PATH, piranha_arguments::PiranhaArgumentsBuilder},
+  models::{
+    default_configs::{JAVA, UNUSED_CODE_PATH},
+    filter::Filter,
+    language::PiranhaLanguage,
+    piranha_arguments::PiranhaArgumentsBuilder,
+  },
   utilities::eq_without_whitespace,
 };
 
@@ -264,4 +271,79 @@ fn test_get_edit_for_context_negative() {
     source_code_unit.get_edit_for_context(29_usize, 33_usize, &mut rule_store, &vec![rule]);
   // let edit = rule.get_edit(&source_code_unit, &mut rule_store, node, true);
   assert!(edit.is_none());
+}
+
+// Tests for not_enclosing_node
+fn run_test_satisfies_filters_not_enclosing_node(
+  filter: Filter, // Replace with the filter to test
+  assertion: fn(bool) -> bool,
+) {
+  let _rule = piranha_rule! {
+    name= "test",
+    query= "(
+      ((local_variable_declaration
+                      declarator: (variable_declarator
+                                          name: (_) @variable_name
+                                          )) @variable_declaration)
+      )",
+    replace_node= "variable_declaration",
+    replace= "",
+    filters= [filter,]
+  };
+  let rule = InstantiatedRule::new(&_rule, &HashMap::new());
+  let source_code = "class Test {
+      public void foobar(){
+        if (isFlagTreated) {
+          int testNumber = 0;
+        }
+       }
+      }";
+
+  let mut rule_store = RuleStore::default();
+  let java = PiranhaLanguage::from(JAVA);
+  let mut parser = java.parser();
+  let piranha_args = PiranhaArgumentsBuilder::default()
+    .path_to_codebase(UNUSED_CODE_PATH.to_string())
+    .language(java)
+    .build();
+  let source_code_unit = SourceCodeUnit::new(
+    &mut parser,
+    source_code.to_string(),
+    &HashMap::new(),
+    PathBuf::new().as_path(),
+    &piranha_args,
+  );
+
+  let start = Point::new(3, 10);
+  let end = Point::new(3, 29);
+  let node = &source_code_unit
+    .root_node()
+    .descendant_for_point_range(start, end)
+    .unwrap();
+
+  let map: HashMap<String, String> = HashMap::new();
+  assert!(assertion(source_code_unit.is_satisfied(
+    *node,
+    &rule,
+    &map,
+    &mut rule_store,
+  )));
+}
+
+#[test]
+fn test_satisfies_filter_not_enclosing_node_positive() {
+  run_test_satisfies_filters_not_enclosing_node(
+    filter! {,
+    not_enclosing_node = "(if_statement) @if_stmt"},
+    |result| !result,
+  );
+}
+
+#[test]
+fn test_satisfies_filter_not_enclosing_node_negative() {
+  run_test_satisfies_filters_not_enclosing_node(
+    filter! {,
+    not_enclosing_node = "(while_statement ) @while"},
+    |result| result,
+  );
 }
