+++ src/models/unit_tests/source_code_unit_test.rs
@@ -13,10 +13,6 @@ Copyright (c) 2023 Uber Technologies, Inc.
 
 use tree_sitter::{Parser, Point};
 
-use crate::models::filter::FilterBuilder;
-
-use crate::models::rule_graph::RuleGraphBuilder;
-use crate::utilities::tree_sitter_utilities::TSQuery;
 use crate::{
   filter,
   models::{
@@ -494,160 +490,4 @@ fn test_contains_no_enclosing_positive() {
     at_least =2},
     |result| result,
   );
-}
-
-// Tests for not_enclosing_node
-fn run_test_satisfies_filters_not_enclosing_node(
-  filter: Filter, // Replace with the filter to test
-  assertion: fn(bool) -> bool,
-) {
-  let _rule = piranha_rule! {
-    name= "test",
-    query= "(
-      ((local_variable_declaration
-                      declarator: (variable_declarator
-                                          name: (_) @variable_name
-                                          )) @variable_declaration)
-      )",
-    replace_node= "variable_declaration",
-    replace= "",
-    filters= [filter,]
-  };
-  let rule = InstantiatedRule::new(&_rule, &HashMap::new());
-  let source_code = "class Test {
-      public void foobar(){
-        if (isFlagTreated) {
-          int testNumber = 0;
-        }
-       }
-      }";
-
-  let mut rule_store = RuleStore::default();
-  let java = get_java_tree_sitter_language();
-  let mut parser = java.parser();
-  let piranha_args = PiranhaArgumentsBuilder::default()
-    .path_to_codebase(UNUSED_CODE_PATH.to_string())
-    .language(java)
-    .build();
-  let source_code_unit = SourceCodeUnit::new(
-    &mut parser,
-    source_code.to_string(),
-    &HashMap::new(),
-    PathBuf::new().as_path(),
-    &piranha_args,
-  );
-
-  let start = Point::new(3, 10);
-  let end = Point::new(3, 29);
-  let node = &source_code_unit
-    .root_node()
-    .descendant_for_point_range(start, end)
-    .unwrap();
-
-  let map: HashMap<String, String> = HashMap::new();
-  assert!(assertion(source_code_unit.is_satisfied(
-    *node,
-    &rule,
-    &map,
-    &mut rule_store,
-  )));
-}
-
-#[test]
-fn test_satisfies_filter_not_enclosing_node_positive() {
-  run_test_satisfies_filters_not_enclosing_node(
-    filter! {,
-    not_enclosing_node = "(if_statement) @if_stmt"},
-    |result| !result,
-  );
-}
-
-#[test]
-fn test_satisfies_filter_not_enclosing_node_negative() {
-  run_test_satisfies_filters_not_enclosing_node(
-    filter! {,
-    not_enclosing_node = "(while_statement ) @while"},
-    |result| result,
-  );
-}
-
-#[test]
-#[should_panic(
-  expected = "Invalid Filter Argument. `at_least` or `at_most` is set, but `contains` is empty !!!"
-)]
-fn test_filter_bad_arg_at_least() {
-  FilterBuilder::default().at_least(2).build();
-}
-
-#[test]
-#[should_panic(
-  expected = "Invalid Filter Argument. `at_least` or `at_most` is set, but `contains` is empty !!!"
-)]
-fn test_filter_bad_arg_at_most() {
-  FilterBuilder::default().at_least(5).build();
-}
-
-#[test]
-#[should_panic(
-  expected = "Invalid Filter Argument. `contains` and `not_contains` cannot be set at the same time !!! Please use two filters instead."
-)]
-fn test_filter_bad_arguments_contains_not_contains() {
-  FilterBuilder::default()
-    .contains(TSQuery::new(String::from("(if_statement) @if_stmt")))
-    .not_contains(vec![TSQuery::new(String::from("(for_statement) @for"))])
-    .build();
-}
-
-#[test]
-#[should_panic(
-  expected = "Invalid Filter Argument. `at_least` should be less than or equal to `at_most` !!!"
-)]
-fn test_filter_bad_range() {
-  FilterBuilder::default()
-    .contains(TSQuery::new(String::from("(if_statement) @if_stmt")))
-    .at_least(5)
-    .at_most(4)
-    .build();
-}
-
-#[test]
-#[should_panic(expected = "Cannot parse")]
-fn test_filter_syntactically_incorrect_contains() {
-  FilterBuilder::default()
-    .contains(TSQuery::new(String::from("(if_statement @if_stmt")))
-    .build();
-}
-
-#[test]
-#[should_panic(expected = "Cannot parse")]
-fn test_filter_syntactically_incorrect_not_contains() {
-  FilterBuilder::default()
-    .not_contains(vec![TSQuery::new(String::from("(if_statement @if_stmt"))])
-    .build();
-}
-
-#[test]
-#[should_panic(expected = "Cannot parse")]
-fn test_filter_syntactically_incorrect_enclosing_node() {
-  FilterBuilder::default()
-    .enclosing_node(TSQuery::new(String::from("(if_statement @if_stmt")))
-    .build();
-}
-
-#[test]
-#[should_panic(expected = "Cannot parse")]
-fn test_filter_syntactically_incorrect_not_enclosing_node() {
-  FilterBuilder::default()
-    .not_enclosing_node(TSQuery::new(String::from("(if_statement @if_stmt")))
-    .build();
-}
-
-#[test]
-#[should_panic(expected = "Cannot parse")]
-fn test_rule_graph_incorrect_query() {
-  RuleGraphBuilder::default()
-    .rules(vec![
-      piranha_rule! {name = "Test rule", query = "(if_statement"},
-    ])
-    .build();
 }
